Extract the generation of a line of empty spaces
This function could be handy elsewhere, and extracting it makes it
easier to write tests for (not a bad idea considering it uses unsafe
code).

The reason for printing a line of spaces as part of rendering the status
bar isn't documented anywhere and may seem a bit strange, so here we
go...

The status bar has a background colour set (unlike the rest of the UI).
If we set the background colour, then print all the status bar contents
we'll end up islands of text with the background colour set, but with
space, which isn't text painted without a background colour.

To work around this we first print the colour codes then paint a full
line of spaces. This correctly paints the whole bar with the appropriate
background colour.

I've tested this on urxvt, however behaviour may possibly vary between
terminals.

---

An alternative to this approach may be to generate a flat string with
spaces between text elements before printing to the terminal.