Fix backspace in prompt
Previously it had pretty weird behaviour and could delete the ':'
(active prompt indicator). This should fix that. To make prompt commands
actually work we'll probably want to create a state machine for the
prompt, which may change the approach we take to rendering - if we
have collected prompt text, we can just clear the line and reprint
it. This may be slower however, so maybe I'll leave it for now.