Introduce CommandPrompt to manage interactions with the command prompt
This is a state machine, tracking inputs and maintaining properties of
the prompt. It doesn't yet handle input parsing but the scaffolding for
passing commands back to the main loop are set up.

The implementation of CommandPrompt may seem a little strange, I think.
It has a few interesting features:

* it consumes itself on each transition - meaning you can't assign the
  result of a transition to a variable and then accidentally trigger
  another state update on the original

* it emits events based on a pull model, assigning the results of the
  last input to it's `last_event` field. I implemented it this way to
  avoid the complexity of working with callbacks. On second thoughts
  maybe it'd make it harder to forget about the event by having
  `CommandPrompt::step` emit a tuple `(CommandPrompt, Event)`.

I'm probably going to pull in `nom`[1] for writing the command parser.
This may be overkill but I want excuses to use it ¯\_(ツ)_/¯

[1]: https://github.com/Geal/nom