Update todo
On narrower terminals I noticed that lines of hex codes would continue
into the status bar. This is kind of unsurprising as I don't do anything
to prevent it. A solution to this might be to implement horizontal
scrolling (possibly repurposing h and l shortcuts, which are currently
bound to changing offset). I prefer this over wrapping lines as hex is
hard enough to read already without having the lines broken.