Improve scrolling logic
Previously it was possible to trigger a panic when trying to scroll with
data which was shorter than the frame. Calculation of `limit_scroll`
would trigger unintentional wrapping (a usize type is used) when
`row_count` is smaller than the frame height - 2.

I've extracted the logic into a separate function and changed the
scrolling behaviour to behave better. Scrolling is essentially disabled
when (scroll is always 0) if there's less than one screen height of data
- otherwise scrolling is allowed up to half a screen past the end of the
data.