Pin top-left byte when adjusting display width
Previously changing the display width would make no change to scroll
position meaning you'd almost always lose track of whatever you were
looking at. This wasn't great.

Now the byte displayed in the top left corner will be anchored in place
when changing the width setting. To ensure the byte is anchored in cases
where width is not being changed in powers of two we also need to make
adjust the offset to ensure the anchored byte remains in the top left
corner.

Changing the offset feels like it could be a little surprising, but
maybe the least bad behaviour.

The initial algorithm I implemented naively increased offset on each
width change that necessitated an offset change, resulting in an
ever-growing offset. To avoid this scroll and offset are coordinated so
that offset never outgrows the display width (in bytes, not columns).

This could be a little unintuitive, and more confusing if you have a
pretty big offset defined, you set the width, and suddenly it's 5, or
some other small number. We'll see how things shake out, I'd be keen
to improve the ergonomics of this.