Ensure cursor location is fixed before printing
When I moved the status bar rendering outside of the individual event
handlers in [commit] it broke the cursor display for the status bar. It
would print the command bar after the individual handlers ran leading
to the cursor sitting on the right hand side of the status bar.

This commit fixes that by always setting the current cursor location
before render. This approach is a bit brute-force but it's good enough
for now.

[commit]: 2f4347b7e7434dcc17239a9720f046891d9d692d