Add todo
When pressing <TAB> in the command prompt the cursor location ends up
broken as the way we render assumes that all characters are equal width.

Part of the issue is that we use the 'cursor' field on `CommandPrompt`
as a representation of the cursor location on the screen, but also as
a character index into the current prompt text string. The former
doesn't really work as:
a) tabs and other whitespace characters are often displayed as
   non-1-width characters on the terminal
b) there are plenty of unicode 'characters' which are displayed
   with a width greater than 1 in terminals

I think a relatively nice solution to this could be to repurpose cursor
as simply being an index into the text string (renaming it would be
nice). This leaves the question of how to ensure the cursor is in the
right place when painting the screen. This should be achievable by
ensuring we render the prompt after *everything else* - in which case
we don't need to worry about calculating/guessing display widths.

I have a suspicion this was broken in [commit-a] and became a visible
issue after [commit-b].

[commit-a]: 2f4347b7e7434dcc17239a9720f046891d9d692d
[commit-b]: e5f2a58ebfc7e0fc6153f174cd48abaa7d379385