Ignore unmapped keys in prompt
Previously we were panicking, this was a bit bad, now we just return an
unmodified prompt object. I'm not sure whether we should force
last_event to Update, or maybe introduce a new explicit "no action"
event but it seems to work ok for now and I can't imagine any issues
with this approach. We'll see how it goes.