Extract screen state into a separate module
Previously we were passing a number of related variables to a bunch of
drawing functions. This was a little unwieldy and involved a lot of
display logic being present in the main run loop. By extracting the data
needed to perform display calculations into a struct, and providing
helper methods on that struct we simplify the main loop and have display
logic grouped, (hopefully) making it easier to understand.

This also opens up potential for further ergonomic improvements.