Parameterize status_bar::render on Screen
By using the Screen struct we can avoid passing a number of related
variables. Additionally this commit introduces a couple of extra
functions on screen which allow status bar printing location to
be more easily configured (in the Screen impl).

It's possible that at a later stage drawing functions *may* end up
being defined on Screen.