Use Line for rendering lines of bytes faster
Previously we were telling the terminal to clear every line before
rendering - this tends to result in a lot of flickering in all terminals
I've used. By using Line, which pads text up to the end of the line with
spaces, we can avoid having to clear lines which is nice.

The reason we *needed* to clear lines, but don't anymore, is the case
which occurs towards the end of a file, where you have a line which is
shorter than full length, scrolling up the screen to replace a line
which is full length...

 old code but without clear line:

  1 FF FF FF FF
  2 AA AA AA

 *scroll down*

  0 FF FF FF FF
  1 AA AA AA FF (oh no)

 new code, or old code with clear line:

  1 FF FF FF FF
  2 AA AA AA

  0 FF FF FF FF
  1 AA AA AA

Unfortunately the approach I've taken to implementing Line leads to a
panic when given more bytes than can be formatted into a single row of
the terminal's width. I'm mostly ok with this as the results were
downright unreadable as the lines would wrap. This has to be fixed
by implementing horizontal scrolling, at which point the current
"panic on too many bytes" is no longer an issue.