Slightly improve error handling in main function
Previously we were unwrapping and expecting everywhere, by moving
the error handling code up a level, and returning on failure
it should become easier to handle encountered errors.

Currently, in the case of failure, we're just returning a Box<Error> and
printing it - but this change should make it easier to introduce a
custom error type which should give us more control over how we respond
to different failure cases.