Update CommandMachine to fully clone on .step()
The semantics I was aiming for with this object would be that it always
returned a copy of itself whenever .step() was called. Previously it
did not, now it does!

This should make it a lot easier to use in cases where a struct
owns the prompt, and assigns the result of prompt.step() to itself
during a method call, e.g.

 struct Prompter {
    prompt: CommandPrompt
 }

 impl Prompt {
     fn do_a_thing(&mut self) {
         self.prompt = prompt.step('c');
     }
 }

Alternatively I could make step() mutate the prompt and return nothing,
this may be preferable in terms of performance, and matching expectations.