Pull rendering (mostly) out of the event handler and into Screen
Rendering was heavily using attributes of the Screen struct which
made me feel like it might belong better there than in the handler.

One outstanding wart is that the handler owns the prompt object and has
to pass it to the screen for it to be rendered. Having the Screen own
the prompt, and expose an interface for handler to call might be nicer.
This also applies to the stdout handle.