Let Screen own stdout
The slightly gnarly lifetime specifiers needed for Screen, and even
moreso for EventHandler are needed as we use stdout after giving the
EventHandler a mutable reference to a Screen, then want to get a mutable
reference to stdout from screen.

The prior lifetimes guaranteed screen, the event handler and stdout
lived as long as each other, now we specify that:
 * screen's reference to data ('b) lives at least as long as event
   handler's reference to screen ('a)
 * stdout (T) lives at least as long as event handler's reference to
   screen ('a)

I'm not sure why both screen's reference to data and stdout's lifetime
need to be specified here but sure.